Make elemToSFItem return [SF] not SF
The way SFs and SEGs are represented is not really consistent with
the XML representation. In particular, SFs can be nested inside the
XML and each SF can have a minnum and a maxnum specified. This way
the XML representation can describe if a group of SEGs is
required. For example imagine an SF, some SEGs are required, some
are not. If this SF is included somewhere with a minnum of 0, then
the logical interpretation is if one of the SEGs inside this SF is
present then all the SEGs inside this SF with a minnum > 0 are
required. This could be logically represented by having minnums and
maxnums both on SFs and SEGs.

The representation of SFs and SEGs in the XML allows considerable
more complex constructs. For example, one could represent an SF
with arbitrary deep nesting. However, this is the not really needed
by HBCI.

For now I have chosen a really simple representation which might be
to simple in some cases. minums and maxnums are only available on
SFs and not on SEGs. SEGs are represented as SFs with a single item
and the minnums and maxnums are relaxed when an SF is referenced.
This might be too simplistic in some cases, I'll fix that if it
turns out to be the case.