Change 'fillMsg' input format to nested maps, eliminate 'SF' type
The SF type was redundant in the way it ended up being used, its fields
got merged into the SEG type.

Providing data in nested maps has the advantage that it is easier to
detect if no data for a subpart was provided and that this should thus
be left out. It also provides more structure to the provided data over
the concatenation of keys which was pretty ad-hoc.

This is an intermediate commit, tests still need to be adapted and fixed.